requirements.txt
Django==1.11.20
django-haystack==2.8.1
Whoosh==2.7.4
